start

+++ Il2CppDumper/Il2CppInspector/Il2CppReaderX86.cs
@@ -14,8 +14,8 @@ namespace Il2CppInspector
         public Il2CppReaderX86(IFileFormatReader stream) : base(stream) { }
         public Il2CppReaderX86(IFileFormatReader stream, uint codeRegistration, uint metadataRegistration) : base(stream, codeRegistration, metadataRegistration) { }
 
-        protected override (uint, uint) Search(uint loc, uint globalOffset) {
-            uint funcPtr, metadata, code;
+        protected override (long, long) Search(long loc, long globalOffset) {
+            long funcPtr, metadata, code;
 
             // Variant 1
